requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
allure-pytest==2.13.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1